+++ src/main/java/com/travelsphere/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.travelsphere.enums.UserStatus;
+package com.travelsphere.enums;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
