+++ src/main/java/com/travelsphere/enums/UserStatus.java
@@ -1,0 +1,5 @@
+package com.travelsphere.enums;
+
+public enum UserStatus {
+    ACTIVE, INACTIVE, PENDING
+}
