--- src/main/java/com/travelsphere/enums/UserStatus/UserStatus.java
@@ -1,5 +1,0 @@
-package com.travelsphere.enums.UserStatus;
-
-public enum UserStatus {
-    ACTIVE, INACTIVE
-}
